fix: correct api and test
